+++ Src/OSF/CAIRO-1160/SRC/cairo-lzw.c
@@ -41,7 +41,7 @@ typedef struct _lzw_buf {
 	uchar * data;
 	int data_size;
 	int num_data;
-	uint32_t pending;
+	uint32 pending;
 	uint pending_bits;
 } lzw_buf_t;
 
@@ -110,25 +110,21 @@ static cairo_status_t _lzw_buf_grow(lzw_buf_t * buf)
  *
  * Sets buf->status to either %CAIRO_STATUS_SUCCESS or %CAIRO_STATUS_NO_MEMORY.
  */
-static void _lzw_buf_store_bits(lzw_buf_t * buf, uint16_t value, int num_bits)
+static void _lzw_buf_store_bits(lzw_buf_t * buf, uint16 value, int num_bits)
 {
 	cairo_status_t status;
-
 	assert(value <= (1 << num_bits) - 1);
-
 	if(buf->status)
 		return;
-
 	buf->pending = (buf->pending << num_bits) | value;
 	buf->pending_bits += num_bits;
-
 	while(buf->pending_bits >= 8) {
 		if(buf->num_data >= buf->data_size) {
 			status = _lzw_buf_grow(buf);
 			if(UNLIKELY(status))
 				return;
 		}
-		buf->data[buf->num_data++] = buf->pending >> (buf->pending_bits - 8);
+		buf->data[buf->num_data++] = static_cast<uchar>(buf->pending >> (buf->pending_bits - 8));
 		buf->pending_bits -= 8;
 	}
 }
@@ -143,22 +139,17 @@ static void _lzw_buf_store_bits(lzw_buf_t * buf, uint16_t value, int num_bits)
 static void _lzw_buf_store_pending(lzw_buf_t * buf)
 {
 	cairo_status_t status;
-
 	if(buf->status)
 		return;
-
 	if(buf->pending_bits == 0)
 		return;
-
 	assert(buf->pending_bits < 8);
-
 	if(buf->num_data >= buf->data_size) {
 		status = _lzw_buf_grow(buf);
 		if(UNLIKELY(status))
 			return;
 	}
-
-	buf->data[buf->num_data++] = buf->pending << (8 - buf->pending_bits);
+	buf->data[buf->num_data++] = static_cast<uchar>(buf->pending << (8 - buf->pending_bits));
 	buf->pending_bits = 0;
 }
 
@@ -173,7 +164,7 @@ static void _lzw_buf_store_pending(lzw_buf_t * buf)
  * 12 bits (19 down to  8):	PREV: previous code value in chain
  *  8 bits ( 7 down to  0):	NEXT: next byte value in chain
  */
-typedef uint32_t lzw_symbol_t;
+typedef uint32 lzw_symbol_t;
 
 #define LZW_SYMBOL_SET(sym, prev, next)                 ((sym) = ((prev) << 8)|(next))
 #define LZW_SYMBOL_SET_CODE(sym, code, prev, next)      ((sym) = ((code << 20)|(prev) << 8)|(next))
